Minor fix to CMakeLists.txt to ensure _cvblob.so python extension binary is built in the expected directory.
Without this fix, the binary is output in ./lib, which is not where the make install step expects to find it,
and so it fails.

I believe the LIBRARY_OUTPUT_PATH variable determines where the extension is output during the build stage, so i just commented
this out, as it was only being used to define the CMAKE_LIB_DIRS_CONFIGCMAKE variable.